Update development_projects.py
I made my changes to the code, the part cleaning the data and merge the data. Hopefully the comments make sense. Let me know if the comments in the code makes sense.

I organized the code so that costar data could be run in a loop, and all redfin data could be run in one loop.